yolo-model/requirements.txt
numpy==1.26.4
ultralytics==8.1.0
Pillow